# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ATLAS DataGob API
  version: 0.7.0
  description: >
    Contract-first API for ATLAS DataGob. The API supports governed demand
    intake, prioritization, portfolio workflows, canonical metadata and
    controlled synthetic-data generation for testing and demonstrations.
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service health status.
  /intake/classify:
    post:
      summary: Classify a business demand into a canonical initiative type.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandPayload"
      responses:
        "200":
          description: Classification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResult"
  /intake/policy-architecture-validate:
    post:
      summary: Validate an intake request against policy RAG, canonical architecture and Operative Committee routing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyArchitecturePayload"
      responses:
        "200":
          description: Policy, architecture and committee validation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyArchitectureValidationResult"
  /synthetic-data/generate:
    post:
      summary: Generate synthetic demand records without deleting existing backlog data.
      description: >
        Creates a new batch of controlled synthetic demands and appends them to
        the configured runtime repository. Existing operational records are
        preserved. Generated records include synthetic provenance metadata and
        a batch identifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyntheticDataPayload"
      responses:
        "200":
          description: Synthetic demand batch generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyntheticDataGenerateResponse"
        "400":
          description: Invalid generation request.
  /policies:
    get:
      summary: Return the versioned policy corpus metadata.
      responses:
        "200":
          description: Policy documents available to the local RAG loader.
  /scoring/calculate:
    post:
      summary: Calculate the prioritization score for a demand.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoringPayload"
      responses:
        "200":
          description: Score and priority bucket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoringResult"
  /metadata/domains:
    get:
      summary: Return the canonical domain and subdomain catalog.
      responses:
        "200":
          description: Data domains and stewardship metadata.
  /metadata/data-dictionary:
    get:
      summary: Return the versioned canonical data dictionary.
      responses:
        "200":
          description: Entities, fields, data types, keys and classification labels.
  /metadata/er-model:
    get:
      summary: Return the canonical entity relationship model.
      responses:
        "200":
          description: Entity relationships and cardinalities.
  /metadata/validate:
    get:
      summary: Validate dictionary, domain catalog and ER model consistency.
      responses:
        "200":
          description: Validation status and errors.
components:
  schemas:
    DemandPayload:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
          minLength: 10
        requester_area:
          type: string
          default: unknown
        requester_role:
          type: string
          default: unknown
        domain_hint:
          type: string
          nullable: true
    PolicyArchitecturePayload:
      allOf:
        - $ref: "#/components/schemas/DemandPayload"
        - type: object
          properties:
            target_consumption:
              type: string
              nullable: true
    ClassificationResult:
      type: object
      required: [initiative_type, confidence, rationale, signals]
      properties:
        initiative_type:
          type: string
          enum: [data_engineering, data_governance, machine_learning, agentic_ai, hybrid, unknown]
        confidence:
          type: number
        rationale:
          type: string
        signals:
          type: array
          items:
            type: string
        secondary_types:
          type: array
          items:
            type: string
    PolicyArchitectureValidationResult:
      type: object
      required:
        - structured_request
        - classification
        - architecture
        - policy_matches
        - policy_gaps
        - architecture_gaps
        - operative_committee
        - human_architecture_review_required
        - recommended_next_action
        - committee_summary
      properties:
        structured_request:
          type: object
        classification:
          type: object
        architecture:
          type: object
        policy_matches:
          type: array
          items:
            type: object
        similar_projects:
          type: array
          items:
            type: object
        policy_gaps:
          type: array
          items:
            type: string
        architecture_gaps:
          type: array
          items:
            type: string
        finops_gaps:
          type: array
          items:
            type: string
        operative_committee:
          type: object
        human_architecture_review_required:
          type: boolean
        recommended_next_action:
          type: string
        committee_summary:
          type: string
    SyntheticDataPayload:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 100
          default: 25
        actor:
          type: string
          default: Synthetic Data Operator

    SyntheticDataGenerateResponse:
      type: object
      required:
        - batch_id
        - count
        - total_backlog
        - demands
        - message
      properties:
        batch_id:
          type: string
          description: Unique identifier for the generated synthetic batch.
        count:
          type: integer
          description: Number of synthetic demands generated.
        total_backlog:
          type: integer
          description: Total backlog size after append-only generation.
        demands:
          type: array
          items:
            type: object
        message:
          type: string

    ScoringPayload:
      type: object
      required:
        - business_value
        - strategic_alignment
        - data_readiness
        - technical_feasibility
        - execution_effort
        - risk_control
      properties:
        business_value: {type: integer, minimum: 1, maximum: 5}
        strategic_alignment: {type: integer, minimum: 1, maximum: 5}
        data_readiness: {type: integer, minimum: 1, maximum: 5}
        technical_feasibility: {type: integer, minimum: 1, maximum: 5}
        execution_effort: {type: integer, minimum: 1, maximum: 5}
        risk_control: {type: integer, minimum: 1, maximum: 5}
    ScoringResult:
      type: object
      required: [score, priority, rationale, components]
      properties:
        score:
          type: number
        priority:
          type: string
          enum: [Alta, Media, Backlog, Reformular]
        rationale:
          type: string
        components:
          type: object